requirements.txt
python>=3.6
biopython
click
matplotlib
networkx
numba
numpy
pandas
scikit-learn
scipy
python-xxhash
